Add files via upload
This C program interacts with the user to perform several tasks based on the input provided:

It begins by asking the user if they want to repeat a process for another number. The program expects the user to enter 'Y' or 'y' to continue or 'N' or 'n' to exit.

Upon a valid choice ('Y' or 'y'), the program proceeds to request a positive integer from the user.

It checks if the entered number is within the valid range (between 2 and 1000) and whether it is a prime number. If the number is not within the specified range or not positive, it provides an appropriate error message.

If the number is valid, the program determines and displays whether it is a prime number.

The program then finds and displays all prime numbers between 2 and the entered positive integer.

It also identifies and shows square pair numbers between 1 and the entered number. Square pair numbers are pairs of integers that, when added together, result in a perfect square and, when subtracted, yield another perfect square.

After processing the entered number, the program prompts the user to repeat the process or exit. If the user enters 'N' or 'n', the program terminates. If 'Y' or 'y' is entered, the process begins anew.